+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Irbtools Changelog
 
-## 4.0 (unreleased)
+## 4.0
 
 * Requires IRB 1.6 (Ruby 3.2 bundles IRB 1.6.2)
 * Requires Ruby 3.0